refactored most of station logic into a module. Also hardcoded selected stations WIP
